secrets: Ignoring the mock server secrets if we are not mocking the DCS server.
