+++ src/main/java/uk/gov/di/ipv/atp/dcs/controllers/AtpController.java
@@ -9,19 +9,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import reactor.core.publisher.Mono;
 import uk.gov.di.ipv.atp.dcs.domain.DcsCheckRequestDto;
 import uk.gov.di.ipv.atp.dcs.services.AtpService;
-import uk.gov.di.ipv.atp.dcs.services.mock.MockDcsServerService;
 
 @Controller
 @Slf4j
 public class AtpController {
 
     private final AtpService atpService;
-    private final MockDcsServerService mockServer;
 
     @Autowired
-    public AtpController(AtpService atpService, MockDcsServerService mockServer) {
+    public AtpController(AtpService atpService) {
         this.atpService = atpService;
-        this.mockServer = mockServer;
     }
 
     @PostMapping("/process")
@@ -30,15 +27,6 @@ public class AtpController {
         var attributes = atpService.checkPassportData(dto);
 
         return attributes
-            .map(ResponseEntity::ok)
-            .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
-
-    @PostMapping("/checks/passport")
-    public Mono<ResponseEntity<String>> check(@RequestBody String jws) {
-        var mockedResponse = mockServer.mockDcs(jws);
-
-        return mockedResponse
             .map(ResponseEntity::ok)
             .defaultIfEmpty(ResponseEntity.badRequest().build());
     }
