+++ src/main/java/uk/gov/di/ipv/atp/dcs/services/mock/DcsMockServerEncryptionServiceImpl.java
@@ -1,6 +1,7 @@
 package uk.gov.di.ipv.atp.dcs.services.mock;
 
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.context.annotation.Conditional;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 import uk.gov.di.ipv.atp.dcs.services.AbstractEncryptionService;
@@ -9,6 +10,7 @@ import java.security.Key;
 import java.security.cert.Certificate;
 
 @Service("mock-server-encryption-service")
+@Conditional(value = MockCondition.class)
 public class DcsMockServerEncryptionServiceImpl extends AbstractEncryptionService {
 
     public DcsMockServerEncryptionServiceImpl(
